requirements.txt
manim
manim-voiceover
google-generativeai
python-dotenv
gTTS
numpy
openai-whisper
ffmpeg-python

flask
